[TASK4] -Add TU's - Giving functionalities restar/sumar

+++ src/main/java/com/example/calculadora/model/ResponseDTO.java
@@ -9,7 +9,7 @@ public class ResponseDTO {
 
     @Getter
     @Setter
-    private int result;
+    private Double result;
 
     public ResponseDTO buildResponse(){return this;}
 }
